users/serializers.py
from rest_framework import serializers
from .models import User


class UserSerializer(serializers.Serializer):
    id = serializers.IntegerField(read_only=True)
    user_nickname = serializers.CharField(max_length=50)
