pyproject.toml
[tool.poetry]
name = "django-2024"
version = "0.1.0"
description = ""
authors = ["seo <seo081225@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
